pyproject.toml
[project]
name = "wyoming-gemini-live"
version = "0.1.0"
description = "Gemini Live (native audio) voice gateway with Home Assistant tool control"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [
  { name = "TBD" }
]
dependencies = [
  "google-genai>=0.2.0",
  "wyoming>=1.5.0",
  "numpy>=1.26.0",
  "scipy>=1.10.0",
  "aiohttp>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
wyoming-gemini-live = "wyoming_gemini_live.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
